refacto and add visualization for BetaDiversity
