+++ moonstone/analysis/diversity/beta.py
@@ -1,20 +1,35 @@
+import logging
+import re
 from abc import ABC, abstractmethod
 
-from skbio.diversity import beta_diversity
+import pandas as pd
+import skbio.diversity
 from skbio.stats.distance import DistanceMatrix
 
-from moonstone.core.module_base import BaseModule, BaseDF
+from moonstone.analysis.diversity.base import DiversityBase
+
+logger = logging.getLogger(__name__)
 
 
-class BetaDiversity(BaseModule, BaseDF, ABC):
-    BETA_DIVERSITY_INDEXES_NAME = "beta_index"
+class BetaDiversity(DiversityBase, ABC):
+    DIVERSITY_INDEXES_NAME = "beta_index"
+    DEF_TITLE = "(beta diversity) distribution across the samples"
+
+    def __init__(self, dataframe: pd.DataFrame):
+        super().__init__(dataframe)
+        self.index_name = " ".join(re.findall('[A-Z][^A-Z]*', self.__class__.__name__)).capitalize()
 
     @abstractmethod
-    def compute_beta_diversity(self) -> DistanceMatrix:
+    def compute_beta_diversity(self, df) -> DistanceMatrix:
         """
         method that compute the beta diversity
         """
         pass
+
+    def compute_diversity(self) -> pd.Series:
+        series = self.beta_diversity.to_series()
+        series.name = self.DIVERSITY_INDEXES_NAME
+        return series
 
     @property
     def beta_diversity(self):
@@ -23,16 +38,25 @@ class BetaDiversity(BaseModule, BaseDF, ABC):
         """
         if getattr(self, '_beta_diversity', None) is None:
             self._beta_diversity = self.compute_beta_diversity(self.df)
-            self._beta_diversity.name = self.BETA_DIVERSITY_INDEXES_NAME
         return self._beta_diversity
 
     @property
     def beta_diversity_series(self):
-        return self.beta_diversity.to_series()
+        return self.compute_diversity()
 
     @property
     def beta_diversity_df(self):
         return self.beta_diversity.to_data_frame()
+
+    def _get_grouped_df(self, metadata_series):
+        df_list = []
+        for group in metadata_series.unique():
+            group_df = self.df.loc[:, metadata_series[metadata_series == group].index]
+            beta_div_indexes = self.compute_beta_diversity(group_df).to_series().reset_index(drop=True).to_frame()
+            beta_div_indexes.columns = [self.DIVERSITY_INDEXES_NAME]
+            beta_div_indexes[metadata_series.name] = group
+            df_list.append(beta_div_indexes)
+        return pd.concat(df_list).dropna()
 
 
 class BrayCurtis(BetaDiversity):
@@ -44,4 +68,4 @@ class BrayCurtis(BetaDiversity):
         :param base: logarithm base chosen (NB : for ln, base=math.exp(1))
         """
         # steps to compute the index
-        return beta_diversity("braycurtis", df.transpose(), df.columns)
+        return skbio.diversity.beta_diversity("braycurtis", df.transpose(), df.columns)
