+++ moonstone/core/module_base.py
@@ -3,6 +3,11 @@ from abc import ABC
 
 import pandas as pd
 
+from moonstone.utils.plot import (
+    add_default_titles_to_plotting_options,
+    add_x_to_plotting_options
+)
+
 logger = logging.getLogger(__name__)
 
 
@@ -10,6 +15,16 @@ class BaseModule(ABC):
     """
     Base class for modules that needs both visualization and report_data
     """
+
+    def _handle_plotting_options(self, plotting_options: dict, title: str, xlabel: str, ylabel: str, log_scale: bool):
+        if plotting_options is None:
+            plotting_options = {}
+        if log_scale:
+            plotting_options = add_x_to_plotting_options(plotting_options, 'yaxes', 'type', "log")
+            ylabel = f"{ylabel} (log)"
+        return add_default_titles_to_plotting_options(
+            plotting_options, title, xlabel, ylabel
+        )
 
     def visualize(self, write_file=False):
         """
